Add the HWMON driver for DA9063
[1/3]
This is the hardware monitor driver component for DA9063 PMIC.

The manual measurement allows monitoring of the system voltage
VSYS, the auxiliary channels ADCIN1, ADCIN2 and ADCIN3, VBBAT
measures of the backup battery voltage and a Tjunc value for the
internal junction temperature sensor.

This patch also includes the necessary register changes, especially
the definitions for ADC_CFG (addr=0xC9) to allow settings for the
current sources.

[2/3]
There are small changes to da9063-core.c to allow reading the temperature
trimming offset value from the PMIC register list. This value is passed
to HWMON and is used during its calculations of the PMIC's internal junction
temperature.

There are also changes to the platform data definitions so that values may
be set for the current sources ADCIN[123]_CUR. These are provided so that
they may be configured outside the driver.

[3/3]
This is the documentation that goes along with the DA9063 HWMON driver
and the associated changes to the DA9063 mfd core and platform data.

[RFC V1 1/3] hwmon: da9063: HWMON driver
[RFC V1 2/3] mfd: da9063: Add HWMON dependencies
[RFC V1 3/3] Documentation: hwmon: New information for DA9063

patch is from lkml

(cherry-picked from 5529911b715fb872b8d9dfe6c4d60e491105814d)