watchdog: da9063_wdt: make sure we dont trigger da9063 to often
The triggering of the watchdog with acknowledge to T_WDMIN is not implemented
in the kernel we are using for Evander
back-porting watchdog code was a hassle, so instead added the check directly
to the da9063 driver
The watchdog did reboot with at time of 256 ms, so I increased the time to 512 ms

Triggering the watchdog faster than T_WDMIN=256ms leads to resets of the
DA9063 chip. The datasheet says that the watchdog must only be triggered
in the timeframe T_WDMIN to T_WDMAX. The T_WDMAX is configured in the
driver.

Signed-off-by: Bo Svangård <bo.svangard@flir.se>
(cherry picked from commit df692ec5fe30e74cf4f218b5ac0ec3816cc47852)