gps: add initialization code to gps after runtime suspend
When doing a runtime suspend, the configuration for interrupts from the gps was lost, causing serial data to be
transmitted over the interupt line. This caused a large amount of interrupts on the device.

Now, when coming out of suspend, the initialisation routing is executed

(cherry picked from commit 22ca9f0bb77b7bb08bc4c365c2c455269625ce8a)