imx7ulp-ec201.dtsi: remove onkey wakeup specifier
Setting the wakeup parameter for onkey registers the pf1550 interrupt as a wakeup interrupt and this for some reason results in the pf1550 interrupt handlers not being triggered after a suspend/resume cycle. Might be because of IRQF_NO_SUSPEND and wakeup setting being non compatible. We wake up the system from the M4 triggered on an SNVS interrupt and don't wake up on the pf1550 interrupt.