mxc_viu.c: reset VIU when stopping stream
Fixes bug when stopping and starting stream after changing format.
Depending on timing this would cause IRQs causing the change to fail and
driver to get in a bad state. According to reference manual the format
should not be changed when viu is running, so resetting the VIU before
changing is the correct behavior.

Signed-off-by: Anton Persson <anton.persson@teledyneflir.com>