BTX-3841 Avoid mixup when closing IPU devices
The IPU device driver has a statically allocated list of memory
structures that is shared among all opened files. Each open file
allocates memory structures upon request and adds them to the list. When
a file is closed the memory structures associated to the specific file
are identified from the list and freed. This requires some sort of
mapping between memory structures in the shared list and opened files.

When an IPU file is opened it is assigned the value of the static
variable file_index and this number is then assigned to all memory
allocated by the file. When the file is close the list of memory is
iterated through and all memory with a matching number to the file is
freed. File_index is incremented on open and decremented on close.

This setup is problematic when a lot of IPU devices are opened and
closed at the same time. Consider the following situation:

  1. Open IPU file, (file_index 1)
  2. Open IPU file, (file_index 2)
  3. Close IPU file from 1. above and free memory assigned index
     value 1.
  4. Open IPU file, (file_index 2). We now have two files opened
     with the same index value of 2.
  5. Close IPU file opened at 2. All memory is freed since all
     memory elements in the shared memory list have index value 2.

Avoid this situation which leads to segmentation fault and various
memory failures by not decrementing file_index when closing an IPU
device.