da9063-regulator: Fix initialization of da9063 regulator
This patch replaces the 3422f8ef281b patch in the 3.14 kernel, which added
a FLIR-made handling of standby entries in the dts. Dealing with the
standby dts entries are now in the mainline kernel so use them.

Further, the da9063 regulator driver is broken regarding the information
handed to main regulator-core dts reading functionality. Fix that by
providing the correct information in the match struct.