[CIRRUS-422] dts: Remove fec clock assignments that break bifrost
Newly introduced clock assignments for fec in imx6qdl.dtsi result in
failure to initiate bifrost on *some* Lennox and Evander cameras.  The
initialization fails on attempts to read from PCI registers and typically
produces the following log output:

[ 16.980784] Unhandled fault: imprecise external abort (0x1406) at 0x7f000000
...
[ 17.283601] [<7f001d44>] (read_device_memory [bifrost]) from [<7f002344>] (bifrost_dma_init+0xac/0x14c [bifrost])

Bisecting the commits between the 5.10.9 (working) and the 5.10.35
(failing) kernels revealed that the "assigned-clocks" and
"assigned-clock-rates" setting in LF-3347/MLK-10199
(cc839f4c1f2719b641861b110301dcee25eeccf5) caused the problem.

It is not completely established how this causes a problem for some cameras
but electrical cross-talk on the debug connection may be one reason.