[CIRRUS-784] tlv320aic31xx.c: Fix sound codec initialization
The codec has specific timing requirements on startup and the init routines
have changed a bit in this new kernel. Fixed the timing, making sure the
chip is reset before power is switched on, and that it stays in reset at
least 4ms after power on.

Also update dts by removing now unnecessary delay on the regulator (handled
in code as above) and minor updates on imx6dl-leif.dtsi for the new kernel
(already in imx6dl-evio.dtsi since CIRRUS-745).