[BC-67] bq24298.c: Decrement notify_psy refcount on shutdown
When the system is going down and the imx6_usb_charger is being removed it
calls power_supply_unregister(). This generated an ugly warning in the
kernel log since bq24298 still held a reference to the
imx6_usb_charger. Solve this by bringing down the bq24298 charger correctly
before the imx6_usb_charger by adding a shutdown function that decrements
the refcount to the usb charger.