[BC-82] fusb30x_driver.c: Use late_initcall for driver
Delay the fusb driver such that its dependencies are started first.

The fusb driver used to (sometimes) start before GPIO (pca953x) and/or
the usb charger. This lead to various problems such as the fusb failing to
start charging at boot time because the bq24298 charger was not ready.