Revert "Fix multiple reference issue"
Reverting as we will now enable imx7d and imx6 in ec302 defconfig. imx7d and imx6 socs are needed when enabling SMP. SMP shouldn't be necessary for imx7ulp since it is a single core processor, but we get a boot error without it. SMP has a fallback mechanism for single core processors, so functionally it should be equivalent, except for the overhead of SMP.

This reverts commit d52130a8d01d8bf0e3d51a29b543d29176fb19e0.

Signed-off-by: Rúni Eriksen <runi.eriksen@teledyneflir.com>