eoco: Use IPU internal clock to drive HDMI
The same external pixel clock is used for all IPU DIs. When reconfiguring
LCD/VF, the HDMI pixclk is also changed. The HDMI image may break if
the resulting framerate is too high for the panel. Avoid this by forcing
HDMI to use the IPU internal clock. The new HDMI pixclk yields 63fps on eoco.

[BC-410]