mxc_ipuv3: only register clone_to attribute on first framebuffer device
Only registering the clone_to attribute on the first framebuffer.

The clone_to attribute is used to allow copying of the main framebuffer display data to other framebuffers
Thus, It is not necessary to register the attribute on other framebuffers than the main (fb0), and
It also causes problems when using the attribute on other frame buffers,
So we only register the attribute on the first registred framebuffer!

The clone_to attribute is currently (known to me) only used on the Eoco platform.
It is still activated on other platforms using mxc_ipuv3, so we can choose to use the
method in those platforms if we choose to

Usage of the clone_to attribute:

The clone_to attribute takes two integer arguments
 - the first argument is activate/deactivate the cloning function
 - the second argument is the framebuffer index, On Eowyn, It is
   2 for viewfinder (fb2), and 3 for HDMI framebuffer (fb3)

echo 1 2 > clone_to, activate copying of framebuffer data from framebuffer 0 to framebuffer 2, with
                     rescaling/rotation/color space conversion as needed
echo 0 2 > clone_to, deactivate copying of framebuffer data from framebuffer 0 to framebuffer 2

echo 1 3 > clone_to, activate copying of framebuffer data from framebuffer 0 to framebuffer 3, with
                     rescaling/rotation/color space conversion as needed
echo 0 3 > clone_to, deactivate copying of framebuffer data from framebuffer 0 to framebuffer 3

[EOWYN-4674]