Eowyn, update framebuffer address in ipu_device
P20-165

For ninjage, the address for the ipu overlay copy recording
is delivered by the framebuffer driver

On Eowyn, that address is changed when the LCD panel is closed, and
the record symbol overlay is moved from FB0 to FB2,
Which makes the UI overlay disappear on recorded and streamed video.

Thus we are separating the implementation between Eoco and Ninjago
platforms, so It will work for both platforms.

We are unable to use the same implementation on Ninjago platforms, since
the EOCO platform will have other streams accessing the fb0 framebuffer
running, which the other Ninjago platforms does not...

Signed-off-by: bo svangård <bo.svangard@flir.se>